requirements.txt
requests
beautifulsoup4
dnspython    